requirements.txt
ipython==8.14.0
moviepy==1.0.3
numpy==1.25.2
opencv_python==4.8.0.76
Pillow==10.0.0
Pillow==10.0.0
